# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My the best API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Api authenticate with passport",
                "description": "Api authenticate with passport",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email address",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Api authenticate with passport",
                "description": "Api authenticate with passport",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout Sucessfull"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/students/{student}/addScore/": {
            "post": {
                "tags": [
                    "score"
                ],
                "description": "Groups store",
                "operationId": "1871b52d8d61285f3669f968f65a2bb8",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student id",
                        "required": true
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Score",
                        "required": true
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Subject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score create sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Score"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/students/{student}/deleteScore/": {
            "delete": {
                "tags": [
                    "score"
                ],
                "description": "Score delete",
                "operationId": "2924ea7c67721e3f03dd37a558dfc950",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student_id",
                        "required": true
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Subject id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score delete sucessfull"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/students/{student}/updateScore": {
            "patch": {
                "tags": [
                    "score"
                ],
                "description": "Score Update",
                "operationId": "487320b597967360f8812e94aff57056",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student id",
                        "required": true
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Score",
                        "required": true
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Subject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score update sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Score"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Score": {
                "required": [
                    "id",
                    "student_id",
                    "subject_id"
                ],
                "properties": {
                    "score": {
                        "description": "Score",
                        "type": "number",
                        "example": "5"
                    },
                    "student_id": {
                        "description": "Student",
                        "type": "number",
                        "example": "11"
                    },
                    "subject_id": {
                        "description": "Subject",
                        "type": "number",
                        "example": "13"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}